server: expose backup as a service
