+++ server/src/scrypted-server-main.ts
@@ -384,28 +384,7 @@ async function start(mainFilename: string, options?: {
 
     app.get('/web/component/backup', async (req, res) => {
         try {
-            const backupDbPath = path.join(volumeDir, 'backup.db');
-            await fs.promises.rm(backupDbPath, {
-                recursive: true,
-                force: true,
-            });
-
-            const backupDb = new Level(backupDbPath);
-            await backupDb.open();
-            for await (const [key, value] of db.iterator()) {
-                await backupDb.put(key, value);
-            }
-            await backupDb.close();
-
-            const backupZip = path.join(volumeDir, 'backup.zip');
-            await fs.promises.rm(backupZip, {
-                recursive: true,
-                force: true,
-            });
-
-            const zip = new AdmZip();
-            await zip.addLocalFolderPromise(backupDbPath, {});
-            const zipBuffer = await zip.toBufferPromise();
+            const zipBuffer = await scrypted.backup.createBackup();
             // the file is a normal zip file, but an extension is added to prevent safari, etc, from unzipping it automatically.
             res.header('Content-Disposition', 'attachment; filename="scrypted.zip.backup"')
             res.send(zipBuffer);
